+++ app/components/map/map_component.tsx
@@ -1,13 +1,13 @@
-import { Map } from "~/components/map/map"
-import { TopBar } from "~/components/overview/top_bar";
-import { BottomNav } from "~/components/overview/bottom_nav"
+import { Map_proper } from "~/components/map/map_proper"
+import { TopBar } from "~/components/shared/top_bar";
+import { BottomNav } from "~/components/shared/bottom_nav"
 
 export function MapComponent() {
   return (
     <div className="flex flex-col h-screen w-full">
         <TopBar isLoginPage={false}></TopBar>
       <div className="flex-1">
-        <Map />
+        <Map_proper />
         <BottomNav />
       </div>
     </div>
