+++ app/components/menu/dish_list.tsx
@@ -1,8 +1,8 @@
 import { FilterBar } from "~/components/overview/filter_bar"
-import { TopBar } from "~/components/overview/top_bar"
-import { BottomNav } from "~/components/overview/bottom_nav"
-//import { mockDishes } from "~/dishlist/mock_dishes"
-import { DishInfo } from "~/menu/dish/dish_info_box"
+import { TopBar } from "~/components/shared/top_bar"
+import { BottomNav } from "~/components/shared/bottom_nav"
+//import { mockDishes } from "~/components/menu/dishlist/mock_dishes"
+import { DishInfo } from "~/components/menu/dish/dish_info_box"
 import { useState } from "react";
 import { useGetDailyMenu } from "~/api/menu_service";
