+++ app/components/overview/overview_component.tsx
@@ -1,9 +1,9 @@
 import { FilterBar } from "~/components/overview/filter_bar"
-import { TopBar } from "~/components/overview/top_bar"
+import { TopBar } from "~/components/shared/top_bar"
 import { SearchBar } from "~/components/overview/search_bar"
 import { ServiceSectionStripe } from "~/components/overview/service_section_stripe"
 import { ServiceSection } from "~/components/overview/service_section/service_section"
-import { BottomNav } from "~/components/overview/bottom_nav"
+import { BottomNav } from "~/components/shared/bottom_nav"
 
 import { useGetAllRestaurants } from "~/api/restaurant_service"
 import { useGetAllVendingMachines } from "~/api/vending_machine_service"
