+++ app/routes/login.tsx
@@ -1,6 +1,6 @@
-import { TopBar } from "~/components/overview/top_bar"
-import { BottomNav } from "~/components/overview/bottom_nav"
-import LoginComponent from "~/login/login";
+import { TopBar } from "~/components/shared/top_bar"
+import { BottomNav } from "~/components/shared/bottom_nav"
+import LoginComponent from "~/components/login/login_component";
 
 export default function Login() {
     return <LoginComponent />
