+++ app/routes/map.tsx
@@ -1,4 +1,4 @@
-import { MapComponent } from "~/map/map";
+import { MapComponent } from "~/components/map/map_component";
 
 export default function Map() {
     return <div className="flex items-center justify-center">
