+++ app/routes/menu.tsx
@@ -1,4 +1,4 @@
-import MenuClient from "../menu/menu_client";
+import MenuClient from "~/components/menu/menu_client";
 
 export default function Menu() {
   // server-rendered route that delegates to a client component
