+++ app/routes/profile.tsx
@@ -1,6 +1,6 @@
-import { TopBar } from "~/components/overview/top_bar"
-import { BottomNav } from "~/components/overview/bottom_nav"
-import ProfileComponent from "~/profile/profile";
+import { TopBar } from "~/components/shared/top_bar"
+import { BottomNav } from "~/components/shared/bottom_nav"
+import ProfileComponent from "~/components/profile/profile_component";
 
 export default function Profile() {
     return <ProfileComponent />
